# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IAM Demo API",
    "version": "2.0.0",
    "description": "API REST documentada para la demo IAM/SOC educativa."
  },
  "servers": [
    { "url": "http://localhost:3001/api" }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
...
  },
  "paths": {
...
    "/resources": {
      "get": {
        "summary": "Lista arbol de recursos",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "Recursos" } }
      },
      "post": {
        "summary": "Crea recurso clasificado",
        "security": [{ "cookieAuth": [] }],
        "responses": { "201": { "description": "Recurso creado" } }
      }
    },
...
  }
}